flask_api/requirements.txt
ultralytics
flask
Pillow
torch
